+++ plugins/modules/oneview_firmware_bundle.py
@@ -17,6 +17,7 @@
 
 from __future__ import (absolute_import, division, print_function)
 from datetime import datetime
+from posixpath import basename
 from ansible_collections.hpe.oneview.plugins.module_utils.oneview import OneViewModule
 __metaclass__ = type
 
@@ -120,13 +121,28 @@ class FirmwareBundleModule(OneViewModule):
             return dict(changed=False, msg=self.MSG_ALREADY_PRESENT, ansible_facts=dict(firmware_bundle=self.current_resource.data))
 
     def __add_compsig(self, file_path):
+        file_name = basename(file_path)
 
-        if self.current_resource and self.current_resource.data.get('resourceState') == 'AddFailed':
+        if self.current_resource and self.current_resource.data.get('resourceState') == 'AddFailed' and\
+                self.current_resource.data.get('signatureFileRequired') is False:
 
             self.current_resource = self.resource_client.upload_compsig(
                 file_path)
             return dict(changed=True, msg=self.MSG_ADD_SIG, ansible_facts=dict(compsig=self.current_resource))
-        elif self.current_resource and self.current_resource.data.get('signatureFileRequired') is True and self.current_resource.data.get('resourceState') == 'Created':
+
+        elif self.current_resource and self.current_resource.data.get('resourceState') == 'AddFailed' and\
+                file_name not in self.current_resource.data.get('signatureFileName'):
+
+            self.current_resource = self.resource_client.upload_compsig(
+                file_path)
+            return dict(changed=True, msg=self.MSG_ADD_SIG, ansible_facts=dict(compsig=self.current_resource))
+            
+        elif self.current_resource and self.current_resource.data.get('resourceState') == 'AddFailed' and\
+                file_name in self.current_resource.data.get('signatureFileName'):
+            return dict(changed=False, msg=self.MSG_SIG_ALREADY_PRESENT)
+
+        elif self.current_resource and self.current_resource.data.get('signatureFileRequired') is True and\
+                self.current_resource.data.get('resourceState') == 'Created':
             return dict(changed=False, msg=self.MSG_SIG_ALREADY_PRESENT)
         else:
             return dict(failed=True, msg=self.MSG_HOTFIX_ABSENT)